trie: move _batch back to trie.batch
